Added support for specifying multiple topics per service
With the changes in this commit, a start() call to the ServiceProvider
could include a map of topic names to associated callbacks. Support for
specifying just a single topic and associated callback is preserved.

Signed-off-by : Jeremy Barlow jeremy_barlow@mcafee.com

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or inpart by me and I have the
right to submit it under the open source license indicated in the file:
or

(b) The contribution is based upon previous work that, to the best
of my knowledge, is covered under an appropriate open source license and
I have the right under that license to submit that work with
modifications, whether created in whole or part by me, under the same
open source license (unless I am permitted to submit under a different
license), as indicated in the file; or

(c) The contribution was provided directly to me by some other person
who certified (a), (b) or (c) I have not modified it.

(d) I understand and agree that this project and the contribution are
public and that a record of the contribution (including all personal
information I submit with it, including my sign-off) is maintained
indefinitely and may be redistributed consistent with this project or
the open source license(s) involved.